# Drop (useless?) `is_expected.to compile` examples