--- a/spec/classes/dnat4_spec.rb
+++ b/spec/classes/dnat4_spec.rb
@@ -57,8 +57,6 @@ describe 'nftables' do
           '
         end
 
-        it { is_expected.to compile }
-
         it {
           expect(subject).to contain_concat('nftables-inet-filter-chain-default_fwd').with(
             path: '/etc/nftables/puppet-preflight/inet-filter-chain-default_fwd.nft',