--- a/spec/classes/rules/afs3_callback_spec.rb
+++ b/spec/classes/rules/afs3_callback_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::afs3_callback' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-afs3_callback') }
         it { is_expected.to contain_nftables__rule('default_in-afs3_callback').with_content('ip saddr { 0.0.0.0/0 } udp dport 7001 accept') }
       end
@@ -20,7 +19,6 @@ describe 'nftables::rules::afs3_callback' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-afs3_callback') }
         it { is_expected.to contain_nftables__rule('default_in-afs3_callback').with_content('ip saddr { 192.168.0.0/16, 1.2.3.4 } udp dport 7001 accept') }
       end