--- a/spec/classes/rules/ceph_spec.rb
+++ b/spec/classes/rules/ceph_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::ceph' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-ceph').with_content('tcp dport 6800-7300 accept comment "Accept Ceph OSD, MDS, MGR"') }
       end
     end