--- a/spec/classes/rules/nfs3_spec.rb
+++ b/spec/classes/rules/nfs3_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::nfs3' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-nfs3').with_content('meta l4proto { tcp, udp } th dport nfs accept comment "Accept NFS3"') }
       end
     end