--- a/spec/classes/rules/out/ceph_client_spec.rb
+++ b/spec/classes/rules/out/ceph_client_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::out::ceph_client' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-ceph_client').with_content('tcp dport { 3300, 6789, 6800-7300 } accept comment "Accept Ceph MON, OSD, MDS, MGR"') }
       end
 
@@ -19,7 +18,6 @@ describe 'nftables::rules::out::ceph_client' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-ceph_client').with_content('tcp dport { 3300, 6790, 6800-7300 } accept comment "Accept Ceph MON, OSD, MDS, MGR"') }
       end
     end