--- a/spec/classes/rules/out/imap_spec.rb
+++ b/spec/classes/rules/out/imap_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::out::imap' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-imap').with_content('tcp dport {143, 993} accept') }
       end
     end