--- a/spec/classes/rules/out/kerberos_spec.rb
+++ b/spec/classes/rules/out/kerberos_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::out::kerberos' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-kerberos_udp').with_content('udp dport 88 accept') }
         it { is_expected.to contain_nftables__rule('default_out-kerberos_tcp').with_content('tcp dport 88 accept') }
       end