--- a/spec/classes/rules/out/nfs_spec.rb
+++ b/spec/classes/rules/out/nfs_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::out::nfs' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-nfs4').with_content('tcp dport nfs accept comment "Accept NFS4"') }
       end
     end