--- a/spec/classes/rules/out/pop3_spec.rb
+++ b/spec/classes/rules/out/pop3_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::out::pop3' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_out-pop3').with_content('tcp dport {110, 995} accept') }
       end
     end