--- a/spec/classes/rules/samba_spec.rb
+++ b/spec/classes/rules/samba_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::samba' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-netbios_tcp').with_content('tcp dport {139,445} accept') }
         it { is_expected.to contain_nftables__rule('default_in-netbios_udp').with_content('udp dport {137,138} accept') }
         it { is_expected.not_to contain_nftables__rule('default_in-ctdb').with_content('tcp dport 4379 accept') }
@@ -21,7 +20,6 @@ describe 'nftables::rules::samba' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-netbios_tcp').with_content('tcp dport {139,445} accept') }
         it { is_expected.to contain_nftables__rule('default_in-netbios_udp').with_content('udp dport {137,138} accept') }
         it { is_expected.to contain_nftables__rule('default_in-ctdb').with_content('tcp dport 4379 accept') }