--- a/spec/classes/rules/ssh_spec.rb
+++ b/spec/classes/rules/ssh_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::ssh' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-ssh').with_content('tcp dport {22} accept') }
       end
 
@@ -19,7 +18,6 @@ describe 'nftables::rules::ssh' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-ssh').with_content('tcp dport {55, 60} accept') }
       end
     end