--- a/spec/classes/rules/tor_spec.rb
+++ b/spec/classes/rules/tor_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::tor' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-tor').with_content('tcp dport {9001} accept') }
       end
 
@@ -19,7 +18,6 @@ describe 'nftables::rules::tor' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-tor').with_content('tcp dport {55, 60} accept') }
       end
     end