--- a/spec/classes/rules/wireguard_spec.rb
+++ b/spec/classes/rules/wireguard_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::rules::wireguard' do
       let(:facts) { os_facts }
 
       context 'default options' do
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-wireguard').with_content('udp dport {51820} accept') }
       end
 
@@ -19,7 +18,6 @@ describe 'nftables::rules::wireguard' do
           }
         end
 
-        it { is_expected.to compile }
         it { is_expected.to contain_nftables__rule('default_in-wireguard').with_content('udp dport {55, 60} accept') }
       end
     end