--- a/spec/classes/services/dhcpv6_client_spec.rb
+++ b/spec/classes/services/dhcpv6_client_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::services::dhcpv6_client' do
       let(:facts) { os_facts }
 
       context 'normal behaviour' do
-        it { is_expected.to compile }
         it { is_expected.to contain_class('nftables::rules::dhcpv6_client') }
         it { is_expected.to contain_class('nftables::rules::out::dhcpv6_client') }
       end