--- a/spec/classes/services/openafs_client_spec.rb
+++ b/spec/classes/services/openafs_client_spec.rb
@@ -8,7 +8,6 @@ describe 'nftables::services::openafs_client' do
       let(:facts) { os_facts }
 
       context 'normal behaviour' do
-        it { is_expected.to compile }
         it { is_expected.to contain_class('nftables::rules::afs3_callback') }
         it { is_expected.to contain_class('nftables::rules::out::openafs_client') }
       end