… spec/defines/simplerule_spec.rb