@jay7x 2023-10-28
Below are some unit test benchmarks on my m1-based macbook

Before the change:
9442 examples, 0 failures, 14 pending

- `rake spec_standalone`: Finished in 11 minutes 10 seconds (files took 4.44 seconds to load)
- `rake parallel_spec_standalone`: Took 187 seconds (3:07)

After the change:
8182 examples, 0 failures, 14 pending

- `rake spec_standalone`: Finished in 8 minutes 56 seconds (files took 4.53 seconds to load)
- `rake parallel_spec_standalone`: Took 168 seconds (2:48)