@duritong 2023-10-30
Afair compile does validate relationships, I don't think other tests do that.