@duritong 2023-10-30
At least in the past compile catched issues that were not visible in other tests.