@duritong 2023-10-30
Yeah this is essentially what we are interested in and what compile gives us: https://rspec-puppet.com/documentation/classes/#test-that-the-catalogue-compiles

> It will test that the manifest can be compiled into a catalogue, and that the catalogue has no dependency cycles between resources.

The second part is what the other tests don't give you afair.